move utils/auth.js to services/auth.js 2
